requirements.txt
pydantic-settings==2.0.2
pyserial==3.5
